add notifications

+++ src/actions/handleSoin.ts
@@ -5,9 +5,28 @@ import { revalidatePath } from "next/cache";
 
 export const deleteSoin = async (id: string) => {
   const supabase = createClient();
+  const { data: user } = await supabase.auth.getUser();
+  const username = user.user?.user_metadata?.username;
+
+  const matchedUser = await supabase
+    .from("soins")
+    .select()
+    .eq("user", username)
+    .eq("id", id)
+    .single();
+  if (!matchedUser.data) {
+    return {
+      success: false,
+      message: "Tu n'as pas le droit de supprimer ceci!",
+    };
+  }
+
   await supabase.from("soins").delete().eq("id", id);
 
   revalidatePath("/dashboard");
+  return {
+    success: true,
+  };
 };
 export const handleTaxi = async (id: string) => {
   const supabase = createClient();
@@ -17,10 +36,26 @@ export const handleTaxi = async (id: string) => {
 export const handlePaye = async (id: string) => {
   const supabase = createClient();
   const { data: user } = await supabase.auth.getUser();
+  const username = user.user?.user_metadata?.username;
+  const { data: alreadyPaid } = await supabase
+    .from("soins")
+    .select("paidBy")
+    .eq("id", id)
+    .single();
+  if (alreadyPaid?.paidBy !== "") {
+    return {
+      success: false,
+      message: "Déjà payé!",
+    };
+  }
+
   await supabase
     .from("soins")
-    .update({ paidBy: user.user?.user_metadata?.username })
+    .update({ paidBy: username })
     .eq("id", id)
     .select();
   revalidatePath("/dashboard");
+  return {
+    success: true,
+  };
 };
